+++ dposlib/network/tlwf.net
@@ -1,7 +1,7 @@
 {
 	"timeout": 10,
 	"begintime": [2016, 5, 24, 17],
-	"familly": "lisk",
+	"familly": "lisk.v09",
 	"maxlimit": 1000,
 	"maxvotepertx": 25,
 	"maxvote": 51,
