+++ setup.py
@@ -30,7 +30,8 @@ kw = {
 		"dposlib.ark",
 		"dposlib.ark.v1",
 		"dposlib.ark.v2",
-		"dposlib.lisk",
+		"dposlib.lisk.v09",
+		"dposlib.lisk.v10",
 		"dposlib.blockchain"
 	],
 	"install_requires": [
